with_timer_in_different_thread
