+++ lumen_runtime/src/otp/erlang/cancel_timer_1/test.rs
@@ -1,12 +1,9 @@
 mod with_local_reference;
 
-use liblumen_alloc::erts::process::Process;
 use liblumen_alloc::erts::term::prelude::*;
 
-use crate::otp::erlang;
 use crate::otp::erlang::cancel_timer_1::native;
-use crate::test::{has_message, timeout_message};
-use crate::{process, timer};
+use crate::timer;
 
 #[test]
 fn without_reference_errors_badarg() {
