+++ lumen_runtime/src/otp/erlang/cancel_timer_1/test/with_local_reference/with_timer/in_different_thread.rs
@@ -3,15 +3,15 @@ use super::*;
 use crate::test::*;
 
 #[test]
+// flaky
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
 
         let message = Atom::str_to_term("different");
         let timeout_message = timeout_message(timer_reference, message, process);
 
-        // flaky
         assert!(!has_message(process, timeout_message));
 
         let milliseconds_remaining =
@@ -35,7 +35,7 @@ fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_mess
 
 #[test]
 fn with_timeout_returns_false_after_timeout_message_was_sent() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
         timeout_after_half(milliseconds, barrier);
