+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test/with_reference_timer_reference/with_list_options/with_async_false/with_info_true/with_local_reference/with_timer/in_different_thread.rs
@@ -5,7 +5,7 @@ use crate::test::*;
 #[test]
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
 
         let message = Atom::str_to_term("different");
@@ -41,7 +41,7 @@ fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_mess
 
 #[test]
 fn with_timeout_returns_false_after_timeout_message_was_sent() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
         timeout_after_half(milliseconds, barrier);
