+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test/with_reference_timer_reference/with_list_options/with_async_true/with_info_true/with_local_reference/with_timer/in_different_thread.rs
@@ -1,15 +1,11 @@
 use super::*;
 
-use std::sync::Barrier;
-use std::thread;
-use std::time::Duration;
-
-use crate::test::with_timer;
+use crate::test::*;
 
 #[test]
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
 
         let message = Atom::str_to_term("different");
@@ -66,7 +62,7 @@ fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_mess
 
 #[test]
 fn with_timeout_returns_false_after_timeout_message_was_sent() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
         timeout_after_half(milliseconds, barrier);
 
@@ -90,10 +86,4 @@ fn with_timeout_returns_false_after_timeout_message_was_sent() {
         );
         assert_eq!(receive_message(process), Some(cancel_timer_message));
     });
-}
-
-fn timeout_after_half(milliseconds: Milliseconds, barrier: &Barrier) {
-    thread::sleep(Duration::from_millis(milliseconds / 2 + 1));
-    timer::timeout();
-    barrier.wait();
 }
