+++ lumen_runtime/src/otp/erlang/read_timer_1/test.rs
@@ -1,12 +1,8 @@
 mod with_local_reference;
 
-use liblumen_alloc::erts::process::Process;
 use liblumen_alloc::erts::term::prelude::*;
 
-use crate::otp::erlang;
 use crate::otp::erlang::read_timer_1::native;
-use crate::process;
-use crate::test::{has_message, timeout_message};
 use crate::timer;
 
 #[test]
