+++ lumen_runtime/src/otp/erlang/read_timer_2/test/with_reference/with_list_options/with_async_false/with_timer/in_different_thread.rs
@@ -6,7 +6,7 @@ use crate::test::*;
 // flaky
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
 
         let message = Atom::str_to_term("different");
@@ -42,7 +42,7 @@ fn without_timeout_returns_milliseconds_remaining() {
 
 #[test]
 fn with_timeout_returns_false_after_timeout_message_was_sent() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
         timeout_after_half(milliseconds, barrier);
