+++ lumen_runtime/src/otp/erlang/read_timer_2/test/with_reference/with_list_options/with_async_true/with_timer/in_different_thread.rs
@@ -1,15 +1,11 @@
 use super::*;
 
-use std::sync::Barrier;
-use std::thread;
-use std::time::Duration;
-
-use crate::test::with_timer;
+use crate::test::*;
 
 #[test]
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
 
         let message = Atom::str_to_term("different");
@@ -81,7 +77,7 @@ fn without_timeout_returns_milliseconds_remaining() {
 
 #[test]
 fn with_timeout_returns_ok_after_timeout_message_was_sent() {
-    with_timer(|milliseconds, barrier, timer_reference, process| {
+    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
         timeout_after_half(milliseconds, barrier);
         timeout_after_half(milliseconds, barrier);
 
@@ -105,10 +101,4 @@ fn with_timeout_returns_ok_after_timeout_message_was_sent() {
         );
         assert_eq!(receive_message(process), Some(read_timer_message));
     });
-}
-
-fn timeout_after_half(milliseconds: Milliseconds, barrier: &Barrier) {
-    thread::sleep(Duration::from_millis(milliseconds / 2 + 1));
-    timer::timeout();
-    barrier.wait();
 }
